Proprietary and protobuf level 2 compression variants implemented.
