+++ client/client-1.0/requests.json
@@ -1,2 +1,3 @@
 {"request":[{"action":"get","path":"Vehicle/Cabin/Door/Row1/Right/IsOpen","requestId":"232"},{"action":"get","path":"Vehicle.Acceleration.Longitudinal","requestId":"233"},
+{"action":"get","path":"Vehicle/ADAS","filter":{"type":"paths","value":["ABS/*","CruiseControl/Error"]},"requestId":"237"},
 {"action":"subscribe","path":"Vehicle/Cabin/Door/Row1/Right/IsOpen","filter":{"type":"timebased","value":{"period":"3"}},"requestId":"246"}]}
