+++ go.mod
@@ -5,12 +5,12 @@ go 1.16
 //example on how to use replace to point to fork or local path
 //replace github.com/MEAE-GOT/WAII/utils => github.com/MagnusGun/WAII/utils master
 //replace github.com/MEAE-GOT/WAII/utils => ./utils
-replace (
-	github.com/COVESA/vss-tools/binary/go_parser/datamodel => /home/ulfbjorkengren/Proj/covesa/vss-tools/binary/go_parser/datamodel
-	github.com/COVESA/vss-tools/binary/go_parser/parserlib => /home/ulfbjorkengren/Proj/covesa/vss-tools/binary/go_parser/parserlib
-)
+//replace (
+//	github.com/COVESA/vss-tools/binary/go_parser/datamodel => /home/ulfbjorkengren/Proj/covesa/vss-tools/binary/go_parser/datamodel
+//	github.com/COVESA/vss-tools/binary/go_parser/parserlib => /home/ulfbjorkengren/Proj/covesa/vss-tools/binary/go_parser/parserlib
+//)
 
-replace github.com/MEAE-GOT/WAII/protobuf/protoc-out => ./protobuf/protoc-out
+//replace github.com/MEAE-GOT/WAII/protobuf/protoc-out => ./protobuf/protoc-out
 
 require (
 	github.com/COVESA/vss-tools/binary/go_parser/datamodel v0.0.0-20211207094201-7208d48f32b6
